builds a new tuple of form ("users", {bindings})
Before we were naively writing out name-address bindings in their own
individual tuples of the form (name, event, binding). This change
leaves those naive writes, but adds an additional 2-tuple where all
the users bindings are stored in a dictionary. This dictionary can be
accessed quickly by matching on a 2-tuple ("users", None) and returned
to the client.